refactor angular tests
